Refactor integration test configuration and update imports
